Refactor implementation to be single-attempt
