+++ openspec/changes/vite-ssr/specs/ssr/spec.md
@@ -167,10 +167,18 @@ When the client disconnects, sku MUST stop document work for that request.
 
 Dev and production MUST share this behaviour.
 
+Disconnect after `render` has a document ready, including during header writes, MUST abort React and MUST NOT start the HTML body.
+
 #### Scenario: Disconnect before headers
 
 - **WHEN** the client disconnects before HTML headers are committed
 - **THEN** sku aborts the stream and does not write the document body
+
+#### Scenario: Disconnect during header writes
+
+- **WHEN** the client disconnects while sku is writing HTML headers
+- **THEN** sku aborts the React stream
+- **AND** sku does not start the HTML body
 
 #### Scenario: Disconnect before a short-circuit Response is written
 
@@ -190,21 +198,25 @@ Dev and production MUST share this behaviour.
 
 ### Requirement: Render attempts settle once
 
-Each document render attempt MUST settle at most once.
+Each `renderToPipeableStream` call is one attempt.
+An attempt settles at most once by becoming ready to commit or by rejecting.
 
-It settles by resolving with a pipeable result or by rejecting.
-
-When the render `AbortSignal` is already aborted, or aborts before settle, the attempt MUST reject with the abort reason.
-
-The attempt MUST NOT hang unsettled after cancellation.
+`render()` MUST honour the render `AbortSignal` before `query()` and before starting a document attempt.
+When the signal is already aborted, or aborts before the document is ready to commit, `render()` MUST reject with the abort reason.
+It MUST NOT run loaders or actions after the signal has aborted.
+It MUST NOT hang unsettled after cancellation.
 
 Cancellation MUST NOT trigger the ErrorBoundary recovery pass.
+
+A failed ErrorBoundary recovery setup MUST reject `render()`.
+It MUST NOT leave the render promise hanging.
 
 #### Scenario: Already-aborted signal rejects promptly
 
 - **WHEN** render starts with an already-aborted signal
 - **THEN** the render promise rejects with the abort reason
-- **AND** sku does not resolve a pipeable document
+- **AND** sku does not run loaders or actions for that request
+- **AND** sku does not resolve a document ready to commit
 
 #### Scenario: Abort during pending waitForAll does not retry
 
@@ -212,6 +224,38 @@ Cancellation MUST NOT trigger the ErrorBoundary recovery pass.
 - **AND** the render signal aborts
 - **THEN** the render promise rejects with the abort reason
 - **AND** sku does not render an ErrorBoundary recovery pass for that abort
+
+#### Scenario: Recovery setup failure rejects
+
+- **WHEN** the first document attempt fails with a recoverable render error
+- **AND** sku cannot build an ErrorBoundary recovery context
+- **THEN** the render promise rejects with that error
+- **AND** sku does not leave the render promise hanging
+
+### Requirement: Uncommitted document renders have a deadline
+
+Sku MUST abort a document render that stays uncommitted past a sku-owned deadline.
+
+That timeout is not a client disconnect.
+When the client is still connected, the rejection MUST reach Express error handling.
+
+The same deadline MUST abort remaining React work after the body has started so a hung Suspense boundary cannot hold the socket open.
+
+Timeout MUST NOT start an ErrorBoundary recovery pass.
+
+#### Scenario: waitForAll exceeds the deadline
+
+- **WHEN** a `waitForAll` render is still waiting for `onAllReady`
+- **AND** the sku-owned deadline elapses
+- **THEN** the render promise rejects
+- **AND** sku does not render an ErrorBoundary recovery pass for that timeout
+- **AND** sku passes the error to Express error handling while the client is connected
+
+#### Scenario: Deadline after piping starts
+
+- **WHEN** the HTML body has started piping
+- **AND** the sku-owned deadline elapses
+- **THEN** sku aborts the React stream
 
 ### Requirement: Cancellation is not an Express render error
