+++ packages/sku/src/services/vite/ssr/abortReason.ts
@@ -1,0 +1,3 @@
+export const abortReason = (signal: AbortSignal | undefined): unknown =>
+  signal?.reason ??
+  new DOMException('This operation was aborted', 'AbortError');
